Timed loop added and tracks move trains to arrival
Trains now signal their arrival properly and tracks move them to the
arrival list.

There is now a loop that counts the amount of milliseconds it took to
process and limits the FPS to as close to 1fps as it can. Needed to use
C++11 for std::chrono features, so the makefile reflects this.