Added cargo to the station and trains
Added cargo for the stations to fill trains with. The cargo timer was
added which mirrors the check point logic pretty closely. When check
logic is finished, it will go on to cargo logic. Both cargo_logic and
check_logic must be true (finished) before the train will be sent off.